Updated Readme, removed instructions, removed console.logs, moved files for better structure
