+++ src/app.css
@@ -70,26 +70,6 @@ input[type='checkbox'] {
   outline: none;
 }
 
-/*
-.dropdown {
-  position: relative;
-}
-
-select {
-  appearance: none;
-  -webkit-appearance: none;
-  width: 100%;
-  font-size: 1.15rem;
-  padding: 0.675em 1em 0.675em 1em;
-  background-color: #ffffffe1;
-  border: 1px solid #caced1;
-  border-radius: 0.25rem;
-  color: #000;
-  cursor: pointer;
-  margin: 1em 0;
-}
-
-*/
 p span {
   font-size: 16px;
   font-weight: 700;
