+++ src/components/survey-components/data/companyTexts.jsx
