+++ src/components/survey-components/data/levelTexts.jsx
