+++ src/components/survey-components/data/locationTexts.jsx
