+++ src/components/survey-components/data/resulttexts.jsx
